optimization
After performance profiling, this is an important Server Lag spike (Server totally no response for ~3s in every ~30s, even in SSP), because of the function it calls has many for's and IOs, and this kind of file-saving for should be fine in another thread.

After performance profiling ,this reflection here occupies 3% of all the cpu time of Client Thread, resulting ~3% fps drop. So this change makes fps increase by ~3%.

This is a code fragment of ancient version of GC and should be removed.

After performance profiling, this change makes fps gain by 2%.
it is not useful to iterate through all the tileentities in the world in every frame, just renew it once per second is OK.